fix(next.config.js): add "cdn.discordapp.com" to allowed domains fix(route.ts): add validation for comment length in POST route fix(route.ts): add DELETE route for deleting comments fix(route.ts): add eai parameter to GET route for excluding incrementing views feat(comments.tsx): add support for posting and deleting comments
feat(blog): add BlogContent component for displaying blog post content

This commit adds a new file `content.tsx` which contains the `BlogContent` component. This component is responsible for displaying the content of a blog post. It imports various components such as `Markdown`, `LoginWithTwitter`, `Card`, `Avatar`, `CardSpotlight`, `TitleAndSubTitle`, etc. It also imports types related to the blog post and page props.

The `BlogContent` component fetches the blog post data from the API using the `useFetch` hook. It also determines the browser language and sets it in the state. The component handles different scenarios such as when the data is not available, when there is an error, when the user is not logged in and the post is not public, and when the data is available.

When the data is available, the component renders the blog post content along with relevant information such as tags, author, creation date, and

refactor(page.tsx): remove unused imports and variables
feat(page.tsx): add generateMetadata function to generate metadata for blog post page

fix(markdown-element.tsx): import Image from "next/image" to use Next.js Image component
feat(markdown-element.tsx): add support for width and height attributes in image elements
fix(markdown-element.tsx): fix rendering of nested link elements in paragraphs